Replace deprecated Query.get() with Session.get() for SQLAlchemy 2.x compatibility.

+++ server/models.py
@@ -30,7 +30,7 @@ class BakedGood(db.Model, SerializerMixin):
 
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String, unique=True)
-    price = db.Column(db.Integer)
+    price = db.Column(db.Float)
     created_at = db.Column(db.DateTime, server_default=db.func.now())
     updated_at = db.Column(db.DateTime, onupdate=db.func.now())
